+++ .idea/libraries/Maven__com_aventstack_extentreports_3_1_5.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: com.aventstack:extentreports:3.1.5">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/aventstack/extentreports/3.1.5/extentreports-3.1.5.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/com/aventstack/extentreports/3.1.5/extentreports-3.1.5-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/aventstack/extentreports/3.1.5/extentreports-3.1.5-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
