+++ .idea/libraries/Maven__org_apache_poi_poi_scratchpad_3_9.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.apache.poi:poi-scratchpad:3.9">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/poi/poi-scratchpad/3.9/poi-scratchpad-3.9.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/poi/poi-scratchpad/3.9/poi-scratchpad-3.9-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/poi/poi-scratchpad/3.9/poi-scratchpad-3.9-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
